[Tests] Add `natural_aligment` to test_types.hpp
